+++ src/com/bbn/openmap/layer/shape/displayindex/DisplaySpatialIndexImpl.java
@@ -543,7 +543,7 @@ public abstract class DisplaySpatialIndexImpl implements DisplaySpatialIndex {
 
   /**
    * Indexes the list of graphics.
-   * 
+   *
    * @param bounds if not null, add min/max values to them.
    * @return The size of the list
    * @throws IOException On error