+++ src/edu/uclouvain/core/nodus/utils/ModalSplitMethodsLoader.java
@@ -53,7 +53,7 @@ public class ModalSplitMethodsLoader {
 
   /**
    * Returns the available modal split methods.
-   * 
+   *
    * @return A linked list that contains the available modal split methods.
    */
   public static LinkedList<ModalSplitMethod> getAvailableModalSplitMethods() {