+++ changelog.md
@@ -266,4 +266,9 @@ value set to "permit". This is not anymore the case since Java 16. It must there
 
 ## v8.1 - Build20211012
 - Upgrade to Parsii 5.0.1 (handles scientific notation)
+
+## v8.1 - BuildXXXXXXXX
+- Remove support for MacBook touchbar. Not useful and not anymore present on the latest MacBook's
+- Reformat some code to apply latest checkstyle rules.
+